+++ src/azarrot/frontends/openai_frontend.py
@@ -2,7 +2,7 @@ import json
 import logging
 import uuid
 from collections.abc import Generator
-from datetime import datetime, timedelta
+from datetime import datetime
 from typing import Any
 
 from fastapi import APIRouter, FastAPI
@@ -315,18 +315,7 @@ class OpenAIFrontend:
         return resp
 
     def __log_generation_statistics(self, generation_statistics: GenerationStatistics) -> None:
-        time_delta = (generation_statistics.end_time - generation_statistics.start_time) / timedelta(milliseconds=1)
-        ftt = (generation_statistics.first_token_time - generation_statistics.start_time) / timedelta(milliseconds=1)
-
-        self._log.info(
-            "Total tokens: %d (prompt %d, completion %d), first token latency: %d ms, cost %d ms, %.3f tok/s",
-            generation_statistics.prompt_tokens + generation_statistics.completion_tokens,
-            generation_statistics.prompt_tokens,
-            generation_statistics.completion_tokens,
-            ftt,
-            time_delta,
-            (generation_statistics.completion_tokens) / time_delta * 1000,
-        )
+        self._log.info(generation_statistics.to_stats_text())
 
     def __wrap_to_openai_chat_completion_stream(
         self,