+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The following are change log of each versions.
 
 ## 0.4.0 (2024-12-??)
 
+- [BREAKING] All OpenAI endpoints are prefixed with `/openai/`
+- Support Jina and Cohere Rerank API
 - Add sentence-transformers backend
 - Update transformers to 4.46.3
 - [WIP] Support assistant API